Fix install script and loracom bugs found in review
- Restore the executable bit on system-setup.sh and the install-*.sh
  scripts, without which sudo ./system-setup.sh fails with rc=126.
- Order cmdb-bootstrap.service against dev-ttyAMA3.device instead of
  dev-serial0.device, which is a different UART (ttyAMA10 on a CM5).
- Build loracom as the invoking user, so root-owned objects in obj/ and
  bin/ no longer break every later non-root make.
- Symlink /usr/local/bin/loracom, so it resolves from systemd services,
  cron, ssh commands and CMDB_POST_INIT_CMD, none of which read
  /etc/profile.d.
- Scale the DeviceLock wait to 3*(retries+1), so a waiting instance can
  no longer give up with -EBUSY while the holder is still inside its own
  retry budget on a noisy link.
- Cast the frame length bytes to uint32_t before shifting them, a top
  byte of 0x80 or more overflowed a signed int.
- Keep any other aliases sharing the 127.0.1.1 line when syncing the
  hostname into /etc/hosts.
- Validate the composed hostname and its length, not just CMDB_ID, so a
  bad CMDB_HOSTNAME_PREFIX can't be half-applied.
- Skip argv[0] in ArgParser, reject trailing characters and negative
  values for unsigned options, and stop flattening the out-of-range
  message into a generic one.
- Remove the temp files the config.txt and /etc/environment rewrites
  leave behind when they fail mid-write.
- Drop the stale comments claiming the CRC and checksum verification are
  unimplemented, and that obj/ and bin/ are committed.
- Correct the README: the UART device is not configurable, and the lock
  retry count is no longer 2*retries.

+++ firmware_CMDB/loracom/LoRaCom.cpp
@@ -35,7 +35,13 @@ std::optional<LoRaCom::ParsedFrame> LoRaCom::readFrame(uint32_t timeoutMs)
     for (;;) {
         // Try to carve a complete frame out of what we already have.
         while (rxBuf_.size() >= 6) {
-            uint32_t length = rxBuf_[2] | (rxBuf_[3] << 8) | (rxBuf_[4] << 16) | (rxBuf_[5] << 24);
+            // Cast before shifting: a uint8_t promotes to int, so a top byte
+            // of 0x80 or more would overflow the sign bit - undefined
+            // behaviour, not the wrapped value the expression looks like.
+            uint32_t length = static_cast<uint32_t>(rxBuf_[2]) |
+                              (static_cast<uint32_t>(rxBuf_[3]) << 8) |
+                              (static_cast<uint32_t>(rxBuf_[4]) << 16) |
+                              (static_cast<uint32_t>(rxBuf_[5]) << 24);
 
             if (length > MAX_FRAME_PAYLOAD) {
                 // Not a plausible header - we're mid-garbage. Drop one byte and
